Fully working app!! :D I'd only like to clear the activitys after you click cancel or submit. And I'd also like if it was impossible to submit if your order is empty.

+++ app/src/main/java/e/natasja/natasjawezel__pset5/RestoDatabase.java
@@ -71,7 +71,7 @@ public class RestoDatabase extends SQLiteOpenHelper {
         return instance;
     }
 
-    public void addData(String dish, int price) {
+    public void addData(String dish, Float price) {
         SQLiteDatabase db = this.getWritableDatabase();
 
         if ( field_exists(dish) == true) {
